requirements.txt
Django==4.0
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-filter==24.2
